Botometer scores

Get Botometer scores for given user(s)

predict_botometer(
  users,
  token = NULL,
  key = NULL,
  set_key = FALSE,
  parse = TRUE,
  verbose = TRUE
)

Arguments

users
Vector of user IDs or screen names

token
rtweet token. If NULL (default), then rtweet looks for a token path in .Renviron

key
Botometer API key, which you should be able to get from: https://rapidapi.com/OSoMe/api/botometer/details

set_key
Logical indicating whether to set the botometer key as an R environment variable for current future sessions (on the same machine).

parse
Logical indicating whether to parse return information. If TRUE (the default) then a data.table is returned.

verbose
Logical indicating whether to print updates between each call. Default is TRUE.

Value

A data frame (data.table) of user ID, screen name, and botometer score

Examples

if (FALSE) {
## get botometer scores
b <- predict_botometer(c("kearneymw", "netflix_bot"))

## get full information returned by botometer (as response objects)
r <- predict_botometer(c("kearneymw", "netflix_bot"), parse = FALSE)
}